Fix read width of early part

+++ arch/mips/megasoc/prom.c
@@ -6,7 +6,7 @@
 
 void prom_putchar(char c)
 {
-    while( (readl((uint32_t *)(EARLY_PRINTK_UART_BASE + 0x5)) & 0x40) == 0 );
+    while( (readl((uint8_t *)(EARLY_PRINTK_UART_BASE + 0x5)) & 0x40) == 0 );
     TRIVIALMIPS_UART_OUTB(0x00, c & 0xFF);
 }
